third_party/blink/web_tests/external/wpt/cookies/name/name-expected.txt

This is a testharness.js-based test.
[FAIL] Set a nameless cookie (that has an = in its value)
  assert_equals: The cookie was set as expected. expected "test=2" but got ""
[FAIL] Set a nameless cookie (that has multiple ='s in its value)
  assert_equals: The cookie was set as expected. expected "==test=2b" but got ""
Harness: the test ran to completion.
